Moved metrics to their own file and fixed various go-lint errors
