+++ cmd/gcp-controller-manager/app/metrics.go
@@ -1,0 +1,37 @@
+/*
+Copyright 2018 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package app
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+var (
+	csrApprovalStatus = prometheus.NewCounterVec(prometheus.CounterOpts{
+		Name: "csr_approval_count",
+		Help: "Count of approved, denied and ignored CSRs",
+	}, []string{"status", "kind"})
+	csrApprovalLatency = prometheus.NewHistogramVec(prometheus.HistogramOpts{
+		Name: "csr_approval_latencies",
+		Help: "Latency of CSR approver, in seconds",
+	}, []string{"status", "kind"})
+)
+
+func init() {
+	prometheus.MustRegister(csrApprovalStatus)
+	prometheus.MustRegister(csrApprovalLatency)
+}
